[#11570] Move all external service classes to new package (#11577)
* Move all external service classes to new package

* Update design diagram and documents

* Update design document and diagram

* Amend docs based on comments
