+++ src/main/java/teammates/logic/api/LogsProcessor.java
@@ -7,9 +7,9 @@ import teammates.common.datatransfer.FeedbackSessionLogEntry;
 import teammates.common.datatransfer.QueryLogsResults;
 import teammates.common.datatransfer.logs.QueryLogsParams;
 import teammates.common.util.Config;
-import teammates.logic.core.GoogleCloudLoggingService;
-import teammates.logic.core.LocalLoggingService;
-import teammates.logic.core.LogService;
+import teammates.logic.external.GoogleCloudLoggingService;
+import teammates.logic.external.LocalLoggingService;
+import teammates.logic.external.LogService;
 
 /**
  * Handles operations related to logs reading/writing.
