+++ src/main/java/teammates/logic/external/EmptyEmailService.java
@@ -1,4 +1,4 @@
-package teammates.logic.core;
+package teammates.logic.external;
 
 import org.apache.http.HttpStatus;
