+++ src/main/java/teammates/logic/external/GoogleCloudLoggingService.java
@@ -1,4 +1,4 @@
-package teammates.logic.core;
+package teammates.logic.external;
 
 import java.time.Instant;
 import java.util.ArrayList;
