+++ src/main/java/teammates/logic/external/LogService.java
@@ -1,4 +1,4 @@
-package teammates.logic.core;
+package teammates.logic.external;
 
 import java.util.List;
