+++ src/main/java/teammates/logic/external/MailjetService.java
@@ -1,4 +1,4 @@
-package teammates.logic.core;
+package teammates.logic.external;
 
 import org.apache.http.HttpStatus;
 import org.json.JSONArray;
