+++ src/main/java/teammates/logic/external/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains the logic of the external dependencies and services.
+ */
+package teammates.logic.external;
