+++ src/test/java/teammates/logic/external/GoogleRecaptchaServiceTest.java
@@ -1,4 +1,4 @@
-package teammates.logic.core;
+package teammates.logic.external;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
