+++ src/test/java/teammates/logic/external/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains test cases for {@link teammates.logic.external} package.
+ */
+package teammates.logic.external;
